Moving more plotting functions to the dedicated module. Adding an extra installation dependencies for the plotting.
